AppDomain refactoring for threads and DLL.
Removed the need to call an .exe for Addins. This was introduced
in the past since we were using SAPBouiCOM.Framework.Application, that
relies on the entry assembly. Since now we have our own implementation
of Framework, we do not need that.

Now from AddinManager we instanciate a B1Application directly from
the Addin appDomain and call the Run method.

There were a problem also using multiple AppDomains, regarding Garbage
Collector and Remoting Proxies, that lead to object timeouts.

To solve this we implemented the Sponsor interface and for each Remote object
we create a Sponsor object for the lifecycle of that instance, eliminating
the timeout error.