External references as embedded resources.
To minimize the number of DLLs that the solution needs to track version, md5sum and
make the manual upgrade an easy step, all external references are now embedded
into Framework.dll.

In order for this to work, it's necessary to register an AppDomain eventHandler for
AssemblyResolve event, and remove the array of system resources in AssemblyManager.

In MicroCore and AddinManager, PrepareForInception must be called after B1Application
startup, otherwise it won´t find any assembly references.