Database Schema modifications
Since all i18n resources are saved in a separate dll file by the .Net
Framework, a new table was added to track that.

In the GA_AO_MODULES table two new fields were created, Installed and AutoUpdate,
that tracks if the module was installed (tables/queries/udos created) and if
an auto update is needed, respectively.

The auto update feature was the default behaviour untill now, since it's really
usefull for debugging. For each new run, it will check if the DLL in the current folder
is newer from the DLL saved into the database, if it is, it will update it.

This is not the default behaviour anymore, otherwise when restoring a customer
database, it would update all versions automatically, and this may cause problems
such as database schema modifications. If the update is needed, the administrator
need to do it manually.