General enhancements to enable i18n implementation
AssemblyManager now supports a new method, AddInIsValid, that
load all needed data into a separated AppDomain and call
CheckAddinConfiguration in AddinService, remotly, in this newly created
AppDomain. This method for now return a text resume of what will be done during addin
installation and also supports the new ZIP container, that contains
i18n resources.

For i18n resources, in i18nServices a new method for checking if the
i18n code is valid was added and in AssemblyDAO now we have a method
for saving this resource into the database.