B1SResource: Support for srf embedded resources.
Now all form XMLs are stored in a separated way, so we can store it
in the same way no matter if it was extracted from a b1s resource project file
or a srf resource form.

All needed operations to register the srf form were also added in the solution.