i18n Implementation
First working implementation of i18n. Still missing Form i18n, but general resources such as Menus are
already working and all thread related configuration regarding i18n is already done.

Since for i18n we need to track all AppDomains Threads and AddinManagers, a code refactoring was done
during this implementation. In the past all startup was done using the Run method (regardless it was
Inception, Addin or first run), becase we needed to run the code in a separate process, since
Business One Application from Framework had some restrictions. Now the Run method is just used
for the first Run, and separate methods were created.

Another change is that, since there is IPC using AppDomains, we do not need IPC using NamedPipes anymore.
This was removed and some new comunications were introduced. Now it's possible to call AddinManager
running in each Addin AppDomain.

Another problem that was faced in this implementation was that when the language is changed, we cannot
rely anymore in the last Form created, because for some reason this is not the behaviour anymore. For
that a new method to get the FormCount was created, so it's possible to use and rely on Forms.GetForm.
Some form creation validations were also added.