I18N implementation working for UDO.
GetSystemFormXML refactored so we do not provide UIAPIRawForm anymore.

In UDOFormBase, if an xml is found, it's considered an update XML adding all itens, except
0_U_E and 1, that are present in our default empty UDO form. This was used so we can create
all i18n stuff and yet use all default event handler present in Business One.

Since this was the final solution for UDO, all UDO specific configuration for form update
was removed in resourceManager. Some duplicate logging was also removed. Since we parse part
of the XML on UDOFormBase, a new GetSystemFormXML that returns an XDocument was implemented.

Added a callback for OnFormLoadBefore on FormEventHandler. Also added an extra check, if
there isn't any event added for this FormUID.