AddinManager: Unload Assembly after addin configuration.
In order to be able to update/start/shutdown addin during runtime, without
the need to restart the Addon, we need to not hold a reference to
the Assembly during runtime.

To do so, with this modification, the addin configuration is done in a
separate AppDomain, so we can unload it.