DBChange backend implementation
Now it`s possible to check if there will be a database change before
addin registration. If there is, AddinManager and AssemblyManager
return a DataTable XML with all structures that will be affected.

For this implementation I also refactory how Model/SAP is implemented.

I moved some stuff from IBOM to IBO, since it makes more sense. A Friendly
description for a IBO was also added to the interface.