BusinessOneDAOImpl: Problem on object update that afects UDO.
There were two problems with the update method.

One is that it was referencing the wrong objects and properties from reflection, possibly
it was based an old version of the AddDIObject.

The second is that when we serialize the object again, it's generating on a wrong format for Business One, since
B1 is very sensitive on XML formats. So we use the same XML that was read previously from B1. This is also
more eficient.