FrameworkInterface: Comunication between AppDomains done by interfaces.
With this implementation now all comunication between appdomains are done
using interfaces defined in FrameworkInterfaces.

With this it will be possible to strong sign a solution and still load it
on Dover, since dover will not rely on the definitions on Framework.dll,
but it will use the definitions on FrameworkInterface.dll, which won't
have it's version changed.

Factory/ContainerManager: Register proxies for FrameworkInterfaces on a different
way so it will be possible to cast it between trusted AppDomains. The way it was done
before, the Castle Proxy interface were referencing the Framework.dll implementation,
causing a cast exception when version differs.

Form/Admin: For some reason the FormProxy does not understand IAddinManager, and it
says it can't solve a dependency for MarshalByRef Object. This has something to do the
way Windsor Castle create proxies. Because of that, we do not use dependency injection
any more for this case.

Sponsor: Moved to FrameworkInterface, becase a reference for the Sponsor
object is passed between AppDomains, so it must point to
FrameworkInterface, since this DLL won't have it's version changed.

AddinManager: Reference for Framework by name and not by ExecutingAssembly, since the version may differ
between signed assemblies.