License check
Included license check on Addin instalation.
Included license check on Addin AppDomain, since if someone changes Dover
DLL with a Dover that do not check license, we make sure it will fail on
the last stage.

Is is good to remember that this will only work if you´re signing your
assemblies, so they are all have strong names.