Thread synchronization refactoring
Use ManualResetEvent instead of Windows Forms Run() and Exit() api.

This makes the code easier to understand, maintain and debug, since everytime we
notify a an exit, we're notifying an specific object that is hold just by
one AppDomain. This way it's clear which AppDomain is being unloaded.

Needed to update this due to thread issues in UnitTests.