Namespace support.
All addins are loaded now in
%APPDATA%\Roaming\Dover\Server-Database\addins\Namespace

This now prevents naming conflicts.

Previously the addin name was used as Key in a lot of places. Replaced all
by Code.

Created new tests to detect those errors.

Feature of Changelog is not working so it was delected. It is now a
missing feature and will be on my issues log.